+++ src/lang/ADempiere/en/businessPartner.js
@@ -9,14 +9,15 @@
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
 const businessPartner = {
+  edit: 'Edit',
   notFound: 'Business partner not found.',
   emptyBusinessPartner: 'Use the filters to search for a business partner by Code, Name, Email and Phone Number',
   searchWithEnter: 'Enter a value to search for a Business Partners'
