+++ src/lang/ADempiere/en/component/date.js
@@ -9,15 +9,22 @@
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-import sequenceSort from './sequenceSort'
-
-export default {
-  sequenceSort
+const date = {
+  endDate: 'End Date',
+  startDate: 'Start Date',
+  Today: 'Today',
+  Yesterday: 'Yesterday',
+  Week: 'Last Week',
+  currentWeek: 'Current Week',
+  LastMonth: 'Last Month',
+  CurrentMonth: 'Current Month'
 }
+
+export default date
