+++ src/lang/ADempiere/en/component/index.js
@@ -9,19 +9,17 @@
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-const sequenceSort = {
-  available: 'Disponibles',
-  disableWithSearch: 'Si hay una búsqueda la funcionalidad arrastrar y soltar se desabilita',
-  sequence: 'Secuencia',
-  updateSequencesSuccessfully: 'Secuencias actualizadas exitosamente',
-  withoutChanges: 'Secuencias/Ordenado sin cambios'
-}
+import date from './date'
+import sequenceSort from './sequenceSort'
 
-export default sequenceSort
+export default {
+  date,
+  sequenceSort
+}
