+++ src/lang/ADempiere/en/form/index.js
@@ -16,11 +16,12 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-const businessPartner = {
-  edit: 'Editar',
-  notFound: 'Socio de negocio no encontrado.',
-  emptyBusinessPartner: 'Utilice los filtros para realizar la búsqueda de socio de negocio según su Código, Nombre, Email y Teléfono',
-  searchWithEnter: 'Introduzca un valor para buscar Socios de Negocio'
-}
+import priceChecking from './priceChecking'
+import productInfo from './productInfo'
+import workflowActivity from './workflowActivity'
 
-export default businessPartner
+export default {
+  priceChecking,
+  productInfo,
+  workflowActivity
+}
