+++ src/lang/ADempiere/en/page/login.js
@@ -1,0 +1,42 @@
+/**
+ * ADempiere-Vue (Frontend) for ADempiere ERP & CRM Smart Business Solution
+ * Copyright (C) 2017-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
+ * Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com https://github.com/EdwinBetanc0urt
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+const login = {
+  noValidUser: 'Please enter the correct user name',
+  noValidPassword: 'The password can not be empty',
+  invalidLogin: 'User Id or Password invalid',
+  unexpectedError: 'An unexpected error has occurred, try again',
+  capsLock: 'Caps lock is On',
+  submit: 'Submit',
+  name: 'Name',
+  lastName: 'Last name',
+  eMail: 'Email',
+  userName: 'User name',
+  userNameOrEmail: 'User name or Email',
+  userEnrollment: 'Enrollment user',
+  userEnrollmentSuccessful: 'Successful user enrollment, check your email',
+  passwordResetSendLink: 'The link to reset the password was sent to the email ',
+  passwordConfirmNew: 'New password confirm',
+  invalidToken: 'The token used is invalid.',
+  passwordReset: 'Password Reset',
+  createPassword: 'Create password',
+  passwordResetSuccessful: 'Password reset was successful',
+  passwordAndConfirmNotMatch: 'The new password and your confirmation do not match.'
+}
+
+export default login
