+++ src/lang/ADempiere/en/page/processActivity.js
@@ -16,11 +16,13 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-const businessPartner = {
-  edit: 'Editar',
-  notFound: 'Socio de negocio no encontrado.',
-  emptyBusinessPartner: 'Utilice los filtros para realizar la búsqueda de socio de negocio según su Código, Nombre, Email y Teléfono',
-  searchWithEnter: 'Introduzca un valor para buscar Socios de Negocio'
+const processActivity = {
+  actions: 'Action',
+  description: 'Description',
+  logs: 'Summary',
+  name: 'Name',
+  status: 'Status',
+  zoomIn: 'Zoom in'
 }
 
-export default businessPartner
+export default processActivity
