+++ src/lang/ADempiere/es/component/sequenceSort.js
@@ -16,11 +16,12 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-const businessPartner = {
-  edit: 'Editar',
-  notFound: 'Socio de negocio no encontrado.',
-  emptyBusinessPartner: 'Utilice los filtros para realizar la búsqueda de socio de negocio según su Código, Nombre, Email y Teléfono',
-  searchWithEnter: 'Introduzca un valor para buscar Socios de Negocio'
+const sequenceSort = {
+  available: 'Disponibles',
+  disableWithSearch: 'Si hay una búsqueda la funcionalidad arrastrar y soltar se desabilita',
+  sequence: 'Secuencia',
+  updateSequencesSuccessfully: 'Secuencias actualizadas exitosamente',
+  withoutChanges: 'Secuencias/Ordenado sin cambios'
 }
 
-export default businessPartner
+export default sequenceSort
