+++ src/lang/ADempiere/es/form/priceChecking.js
@@ -16,11 +16,9 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-const businessPartner = {
-  edit: 'Editar',
-  notFound: 'Socio de negocio no encontrado.',
-  emptyBusinessPartner: 'Utilice los filtros para realizar la búsqueda de socio de negocio según su Código, Nombre, Email y Teléfono',
-  searchWithEnter: 'Introduzca un valor para buscar Socios de Negocio'
+const priceChecking = {
+  basePrice: 'Precio Base',
+  productNotFound: 'Producto No Disponible'
 }
 
-export default businessPartner
+export default priceChecking
