+++ src/lang/ADempiere/es/page/index.js
@@ -16,11 +16,14 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-const businessPartner = {
-  edit: 'Editar',
-  notFound: 'Socio de negocio no encontrado.',
-  emptyBusinessPartner: 'Utilice los filtros para realizar la búsqueda de socio de negocio según su Código, Nombre, Email y Teléfono',
-  searchWithEnter: 'Introduzca un valor para buscar Socios de Negocio'
+import login from './login.js'
+import processActivity from './processActivity'
+import settings from './settings.js'
+
+const page = {
+  login,
+  processActivity,
+  settings
 }
 
-export default businessPartner
+export default page
