+++ src/lang/ADempiere/es/page/login.js
@@ -1,0 +1,42 @@
+/**
+ * ADempiere-Vue (Frontend) for ADempiere ERP & CRM Smart Business Solution
+ * Copyright (C) 2017-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
+ * Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com https://github.com/EdwinBetanc0urt
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+const login = {
+  noValidUser: 'Por favor ingrese el nombre de usuario correcto',
+  noValidPassword: 'La contraseña no puede estar vacía',
+  invalidLogin: 'ID de usuario o contraseña no válida',
+  unexpectedError: 'Se ha producido un error inesperado, intente nuevamente',
+  capsLock: 'Bloqueo de mayúsculas activado',
+  submit: 'Enviar',
+  name: 'Nombre',
+  lastName: 'Apellido',
+  eMail: 'Correo electronico',
+  userName: 'Usuario',
+  userNameOrEmail: 'Usuario o Correo',
+  userEnrollment: 'Registrar usuario',
+  userEnrollmentSuccessful: 'Registro de usuario exitoso, verifique su correo electronico',
+  passwordResetSendLink: 'El enlace para reiniciar la contraseña se envio al correo ',
+  passwordConfirmNew: 'Confirmación de nueva contraseña',
+  invalidToken: 'El token utilizado es inválido.',
+  passwordReset: 'Reiniciar contraseña',
+  createPassword: 'Crear contraseña',
+  passwordResetSuccessful: 'El reinicio de contraseña se realizo correctamente',
+  passwordAndConfirmNotMatch: 'La contraseña nueva y su confirmación no coinciden.'
+}
+
+export default login
