+++ src/lang/es.js
@@ -163,17 +163,7 @@ export default {
     refresh: 'Actualizar',
     close: 'Cerrar',
     closeOthers: 'Cerrar otros',
-    closeAll: 'Cerrar todos',
-    table: {
-      label: 'Cerrar Ventana',
-      tab: 'Pestaña',
-      field: 'Campo',
-      value: 'Valor'
-    },
-    buttonClose: {
-      discardChanges: 'Descartar Cambios',
-      returnToWindow: 'Regresar a la Ventana'
-    }
+    closeAll: 'Cerrar todos'
   },
   settings: {
     title: 'Page style setting',
