+++ src/store/modules/ADempiere/formDefinition.js
@@ -1,8 +1,30 @@
-import { requestForm } from '@/api/ADempiere/dictionary/form'
-import { showMessage } from '@/utils/ADempiere/notification'
-import { isEmptyValue } from '@/utils/ADempiere/valueUtils'
+/**
+ * ADempiere-Vue (Frontend) for ADempiere ERP & CRM Smart Business Solution
+ * Copyright (C) 2017-Present E.R.P. Consultores y Asociados, C.A. www.erpya.com
+ * Contributor(s): Edwin Betancourt EdwinBetanc0urt@outlook.com https://github.com/EdwinBetanc0urt
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
+
 import router from '@/router'
 import language from '@/lang'
+
+// API Request Methods
+import { requestForm } from '@/api/ADempiere/dictionary/form'
+
+// Utils and Helper Methods
+import { showMessage } from '@/utils/ADempiere/notification'
+import { isEmptyValue } from '@/utils/ADempiere/valueUtils'
 
 const form = {
   state: {
@@ -76,7 +98,7 @@ const form = {
             }, () => {})
             dispatch('tagsView/delView', routeToDelete)
             showMessage({
-              message: language.t('login.unexpectedError'),
+              message: language.t('page.login.unexpectedError'),
               type: 'error'
             })
             console.warn(`Dictionary form - Error ${error.code}: ${error.message}.`)
