+++ src/views/ADempiere/ProcessActivity/modeMobile.vue
@@ -62,7 +62,7 @@
                         </el-dropdown-item>
 
                         <el-dropdown-item :command="{ ...activity, command: 'zoomIn' }">
-                          {{ $t('table.ProcessActivity.zoomIn') }}
+                          {{ $t('page.processActivity.zoomIn') }}
                         </el-dropdown-item>
 
                         <!-- TODO: add more actions -->
@@ -78,7 +78,7 @@
                       :inline="true"
                       @submit.native.prevent="notSubmitForm"
                     >
-                      <el-form-item :label="$t('table.ProcessActivity.Logs')" style="margin-bottom: 0;">
+                      <el-form-item :label="$t('page.processActivity.logs')" style="margin-bottom: 0;">
 
                         <span v-if="activity.isReport && !isEmptyValue(activity.summary)">
                           <ul>
@@ -163,7 +163,7 @@
                         </el-dropdown-item>
 
                         <el-dropdown-item :command="{ ...activity, command: 'zoomIn' }">
-                          {{ $t('table.ProcessActivity.zoomIn') }}
+                          {{ $t('page.processActivity.zoomIn') }}
                         </el-dropdown-item>
 
                         <!-- TODO: add more actions -->
@@ -174,7 +174,7 @@
                 <el-collapse-transition>
                   <div v-show="(currentKey === index)">
                     <el-form label-position="top" style="padding-left: 1%;" :inline="true">
-                      <el-form-item :label="$t('table.ProcessActivity.Logs')" style="margin-bottom: 0;">
+                      <el-form-item :label="$t('page.processActivity.logs')" style="margin-bottom: 0;">
 
                         <span v-if="activity.isReport && !isEmptyValue(activity.summary)">
                           <ul>
@@ -261,7 +261,7 @@
                         </el-dropdown-item>
 
                         <el-dropdown-item :command="{ ...activity, command: 'zoomIn' }">
-                          {{ $t('table.ProcessActivity.zoomIn') }}
+                          {{ $t('page.processActivity.zoomIn') }}
                         </el-dropdown-item>
 
                         <!-- TODO: add more actions -->
@@ -272,7 +272,7 @@
                 <el-collapse-transition>
                   <div v-show="(currentKey === index)">
                     <el-form label-position="top" style="padding-left: 1%;" :inline="true">
-                      <el-form-item :label="$t('table.ProcessActivity.Logs')" style="margin-bottom: 0;">
+                      <el-form-item :label="$t('page.processActivity.logs')" style="margin-bottom: 0;">
 
                         <span v-if="activity.isReport && !isEmptyValue(activity.summary)">
                           <ul>
